requirements.txt
Django==5.1.7
djangorestframework==3.15.2
python-dotenv==1.1.0
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
django-filter==25.1